+++ apps/test-suite/screens/SelectScreen.js
@@ -7,10 +7,11 @@ import {
   View,
   Text,
   Platform,
+  Alert,
   Button,
-  StatusBar,
   PixelRatio,
 } from 'react-native';
+import { getTestModules } from '../TestUtils';
 
 class ListItem extends React.Component {
   onPress = () => {
@@ -44,22 +45,24 @@ class ListItem extends React.Component {
 export default class SelectScreen extends React.PureComponent {
   constructor(props) {
     super(props);
-    const initialValues = this.props.data.map(item => [item.name, false]);
 
+    this.modules = getTestModules();
     this.state = {
-      selected: new Map(initialValues),
-      allSelected: false,
+      selected: new Set(),
     };
-    this._selectAll = this._selectAll.bind(this);
-    this._getSelected = this._getSelected.bind(this);
   }
+
+  static navigationOptions = {
+    title: 'Test Suite',
+  };
 
   _keyExtractor = item => item.name;
 
   _onPressItem = id => {
     this.setState(state => {
-      const selected = new Map(state.selected);
-      selected.set(id, !selected.get(id)); // toggle
+      const selected = new Set(state.selected);
+      if (selected.has(id)) selected.delete(id);
+      else selected.add(id);
       return { selected };
     });
   };
@@ -68,49 +71,53 @@ export default class SelectScreen extends React.PureComponent {
     <ListItem
       id={item.name}
       onPressItem={this._onPressItem}
-      selected={this.state.selected.get(item.name)}
+      selected={this.state.selected.has(item.name)}
       title={item.name}
     />
   );
 
   _selectAll = () => {
     this.setState(state => {
-      const selected = new Map(state.selected);
-      for (const key of selected.keys()) {
-        selected.set(key, !state.allSelected);
+      if (state.selected.size === this.modules.length) {
+        return { selected: new Set() };
       }
-      return { selected, allSelected: !state.allSelected };
+      return { selected: new Set(this.modules.map(this._keyExtractor)) };
     });
   };
 
   _getSelected = () => {
     const selected = this.state.selected;
-    let selectedTests = [];
-    for (const test of selected.keys()) {
-      if (selected.get(test)) {
-        selectedTests.push(test);
-      }
+    const selectedModules = this.modules.filter(m => selected.has(m.name));
+    return selectedModules;
+  };
+
+  _navigateToTests = () => {
+    const selected = this._getSelected();
+    if (selected.length === 0) {
+      Alert.alert('Cannot Run Tests', 'You must select at least one test to run.');
+    } else {
+      this.props.navigation.navigate('RunTests', { selected });
+      this.setState({ selected: new Set() });
     }
-    return new Set(selectedTests);
   };
 
   render() {
-    const buttonTitle = this.state.allSelected ? 'Deselect All' : 'Select All';
+    const allSelected = this.state.selected.size === this.modules.length;
+    const buttonTitle = allSelected ? 'Deselect All' : 'Select All';
     return (
       <View style={styles.mainContainer}>
         <FlatList
-          data={this.props.data}
+          data={this.modules}
           extraData={this.state}
           keyExtractor={this._keyExtractor}
           renderItem={this._renderItem}
-          style={styles.flatList}
         />
         <View style={styles.buttonRow}>
           <View style={styles.buttonContainer}>
             <Button title={buttonTitle} onPress={this._selectAll} />
           </View>
           <View style={styles.buttonContainer}>
-            <Button title="Run Tests" onPress={() => this.props.runTests(this._getSelected())} />
+            <Button title="Run Tests" onPress={this._navigateToTests} />
           </View>
         </View>
       </View>
@@ -121,9 +128,6 @@ export default class SelectScreen extends React.PureComponent {
 const styles = StyleSheet.create({
   mainContainer: {
     flex: 1,
-  },
-  flatList: {
-    marginTop: StatusBar.currentHeight,
   },
   listItem: {
     paddingHorizontal: 10,
